EMS-746 code fixes
